Implemented eddy-current/Joule-losses for magneticEdge PDE, now the coupling between EM-Heat can be performed (at least for the linear case) A simple testsuite example (no verification yet) will be added in a few minutes
In 1 out of 5 ctest executions, testcase ElTh_Static had a tolerance of 1.1E-6 and therefore failed. This will be further investigated (tested with gcc-4.8.5).

+++ source/OLAS/multigrid/topology.cc
@@ -635,7 +635,7 @@ GetNumInterpolatedPoints( StdVector<UInt>& sizes ) const
     Integer totalNumPoints = 0;
 
     // run over all fine grid points
-    for( Integer i = 0; i < GetSizeh(); i++ ) {
+    for( Integer i = 0; i < (Integer)GetSizeh(); i++ ) {
         // only coarse points interpolate other points
         if( IsCPoint(i) ) {
             // initialize the counter with 1, because every coarse
